chore: replace message dialog with scaffold messenger
