+++ lib/widgets/text_input_widget.dart
@@ -1,0 +1,36 @@
+import 'package:final_project/consts/app_color.dart';
+import 'package:flutter/material.dart';
+
+class CustomTextField extends StatelessWidget {
+  const CustomTextField({
+    super.key,
+    required this.controller,
+    this.hintText = '',
+    this.maxLines,
+  });
+
+  final TextEditingController controller;
+  final String? hintText;
+  final int? maxLines;
+
+  @override
+  Widget build(BuildContext context) {
+    return TextField(
+      controller: controller,
+      style: const TextStyle(
+        color: AppColors.defaultTextColor,
+      ),
+      maxLines: maxLines,
+      decoration: InputDecoration(
+        hintText: hintText,
+        hintStyle: TextStyle(
+          fontSize: 14,
+          color: AppColors.defaultTextColor.withOpacity(0.3),
+        ),
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+        ),
+      ),
+    );
+  }
+}
